+++ rollup.config.js
@@ -1,4 +1,4 @@
-import babel from 'rollup-plugin-babel';
+import babel from '@rollup/plugin-babel';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from 'rollup-plugin-commonjs';
 
@@ -15,7 +15,9 @@ export default [{
     name: 'typesonRegistrySCAReverter'
   },
   plugins: [
-    babel(),
+    babel({
+      babelHelpers: 'bundled'
+    }),
     resolve(),
     commonjs()
   ]
@@ -26,7 +28,9 @@ export default [{
     format: 'es'
   },
   plugins: [
-    babel(),
+    babel({
+      babelHelpers: 'bundled'
+    }),
     resolve({
       mainFields: ['module']
     }),